Const-cast cleanup phase 1: drop redundant "as TYPE" casts on numeric-literal consts
Addresses the pattern flagged during the misc-const-dedup work:
`pub const NAME: TYPE = LITERAL as TYPE;` - the trailing cast is
inert boilerplate left over from c2rust's translation, since a bare
integer literal already infers TYPE from the const's own declared
type (Rust's numeric-literal type inference propagates through
arithmetic too, so this covers multi-term expressions like
`0x80 as i32 + 0x3b as i32` -> `0x80 + 0x3b` as well, not just
single-literal consts).

Scoped narrowly and mechanically: only strips `LITERAL as TYPE` where
LITERAL is a bare hex or decimal integer literal and TYPE textually
matches the enclosing const's own declared type - 302 occurrences
across 46 files (both pub and (in one nested case) implicitly-pub-via-
outer-fn consts). Deliberately excludes char-literal casts
(`'x' as i32` etc) - confirmed by an early failed attempt that char
literals are NOT subject to the same inference (a char literal's type
is always `char`, never auto-coerced by context), so `'x' as i32` is a
real, necessary conversion and must stay. Also excludes casts applied
to anything other than a bare literal (another const's name, a
function call, a differently-typed expression) - those may be doing
real work and need individual judgement, out of scope for this
mechanical pass.

Verified: 0 build errors (lib + bin, learned from the previous extern-C
phase to always check both), warning-diff against base matches
category-by-category (110 = 110) - byte-identical warning positions
since removing mid-line text doesn't shift line numbers. Release build
succeeds; Xvfb smoke test confirms boot, movement, and weapon fire
all work correctly.

+++ rust-doomgeneric/src/d_iwad.rs
@@ -100,7 +100,7 @@ static iwads: [iwad_t; 14] = [
         description: "Strife",
     },
 ];
-pub const MAX_IWAD_DIRS: i32 = 128 as i32;
+pub const MAX_IWAD_DIRS: i32 = 128;
 static mut iwad_dirs_built: bool = false;
 static mut iwad_dirs: Vec<String> = Vec::new();
