+++ rust-doomgeneric/src/doomdef.rs
@@ -3,9 +3,9 @@ pub type pixel_t = u32;
 pub const NULL: *mut ::core::ffi::c_void = ::core::ptr::null_mut::<
     ::core::ffi::c_void,
 >();
-pub const true_0: i32 = 1 as i32;
-pub const false_0: i32 = 0 as i32;
-pub const MAXPLAYERS: i32 = 4 as i32;
-pub const TICRATE: i32 = 35 as i32;
-pub const SCREENWIDTH: i32 = 320 as i32;
-pub const SCREENHEIGHT: i32 = 200 as i32;
+pub const true_0: i32 = 1;
+pub const false_0: i32 = 0;
+pub const MAXPLAYERS: i32 = 4;
+pub const TICRATE: i32 = 35;
+pub const SCREENWIDTH: i32 = 320;
+pub const SCREENHEIGHT: i32 = 200;
