+++ rust-doomgeneric/src/doomgeneric.rs
@@ -8,8 +8,8 @@ use libc::malloc;
 extern "C" {
     fn DG_Init();
 }
-pub const DOOMGENERIC_RESX: i32 = 640 as i32;
-pub const DOOMGENERIC_RESY: i32 = 400 as i32;
+pub const DOOMGENERIC_RESX: i32 = 640;
+pub const DOOMGENERIC_RESY: i32 = 400;
 #[no_mangle]
 pub static mut DG_ScreenBuffer: *mut pixel_t = ::core::ptr::null::<pixel_t>()
     as *mut pixel_t;
