+++ rust-doomgeneric/src/m_misc.rs
@@ -27,8 +27,8 @@ extern "C" {
     fn mkdir(__path: *const ::core::ffi::c_char, __mode: __mode_t) -> i32;
 }
 pub type __mode_t = u32;
-pub const SEEK_END: i32 = 2 as i32;
-pub const EISDIR: i32 = 21 as i32;
+pub const SEEK_END: i32 = 2;
+pub const EISDIR: i32 = 21;
 pub const DIR_SEPARATOR: i32 = '/' as i32;
 pub const DIR_SEPARATOR_S: [::core::ffi::c_char; 2] = unsafe {
     ::core::mem::transmute::<[u8; 2], [::core::ffi::c_char; 2]>(*b"/\0")
